pyproject.toml
[project]
name = "graphcontainer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.4.1",
    "flagembedding>=1.3.5",
    "openai>=1.108.1",
    "pandas>=3.0.0",
    "python-dotenv>=1.2.1",
    "pyct>=0.6.0",
    "pydantic>=2.12.5",
    "pyvis>=0.3.2",
    "sentence-transformers>=5.2.0",
    "tiktoken>=0.12.0",
    "torch>=2.10.0",
    "tqdm>=4.67.1",
    "transformers>=4.57.6",
]
